fix(lint): split cli.rs and cli_tests.rs under the 600-line gate
cli.rs (679) and cli_tests.rs (617) had crept past the repo's 600-line
pre-push gate. Splits the query-command implementations
(cleanup/trigger/status, their JSON rendering, and health probing) into
cli_query.rs, and their tests into cli_query_tests.rs, following the
existing per-concern sibling-module convention. No behavior or test-body
changes.
